requirements.txt
Flask>=2.0
requests>=2.25
python-dotenv # Good practice for managing environment variables (like API keys later)
# Add firebase-admin later for Phase 2